WIP: Saving progress before starting cleanup
