+++ tests/integration.test.js
@@ -1,140 +1,136 @@
 /**
  * Integration tests for PDF utility functions.
  * 
- * These tests use real PDF files to verify that our utility functions work correctly
- * with the actual pdf-lib library, not just the mocks.
+ * These tests verify that our utility functions work correctly with real PDF data
+ * using only the public API (Buffer in, Buffer out) without exposing implementation details.
  */
 
-const fs = require('fs');
-const path = require('path');
-const { PDFDocument } = require('pdf-lib');
-const { appendPdfs, mergePdfs } = require('../src/pdf-utils');
+const { mergePdfs, appendPdfs } = require('../src/pdf-utils');
 
-// Helper function to create simple PDF documents with specified page count
-const createSamplePdf = async (pageCount = 1) => {
-  const pdfDoc = await PDFDocument.create();
-  
-  // Add specified number of pages
-  for (let i = 0; i < pageCount; i++) {
-    const page = pdfDoc.addPage();
-    
-    // Add some content to the page to make it identifiable
-    const { width, height } = page.getSize();
-    page.drawText(`Page ${i+1}`, {
-      x: 50,
-      y: height - 50,
-      size: 24,
-    });
-  }
-  
-  // Convert Uint8Array to Node.js Buffer
-  return Buffer.from(await pdfDoc.save());
+// Simple helper to create a valid PDF buffer with a specific "signature"
+const createTestPdfBuffer = async (identifier) => {
+  // This is a minimal valid PDF structure
+  return Buffer.from(
+    `%PDF-1.7
+1 0 obj<</Type/Catalog/Pages 2 0 R>>endobj
+2 0 obj<</Type/Pages/Count 1/Kids[3 0 R]>>endobj
+3 0 obj<</Type/Page/MediaBox[0 0 612 792]/Contents 4 0 R>>endobj
+4 0 obj<</Length 23>>stream
+BT /F1 12 Tf 100 700 Td (PDF ${identifier}) Tj ET
+endstream
+endobj
+xref
+0 5
+0000000000 65535 f
+0000000010 00000 n
+0000000053 00000 n
+0000000102 00000 n
+0000000169 00000 n
+trailer<</Size 5/Root 1 0 R>>
+startxref
+240
+%%EOF`
+  );
 };
 
-describe('Integration Tests with Real PDFs', () => {
-  describe('appendPdfs', () => {
-    it('should correctly append pages from one PDF to another', async () => {
-      // ARRANGE: Create real PDF documents
-      const sourcePdfBuffer = await createSamplePdf(3); // 3-page PDF
-      const targetPdfDoc = await PDFDocument.create();
-      targetPdfDoc.addPage(); // Target starts with 1 page
-      
-      // ACT: Append the source pages to the target
-      const result = await appendPdfs(sourcePdfBuffer, targetPdfDoc);
-      
-      // ASSERT:
-      expect(result.success).toBe(true);
-      expect(result.pagesAdded).toBe(3);
-      
-      // Verify the target document now has 4 pages (1 original + 3 added)
-      const targetPages = targetPdfDoc.getPageIndices();
-      expect(targetPages.length).toBe(4);
-    });
-    
-    it('should handle empty source PDFs without errors', async () => {
-      // ARRANGE: Create an empty PDF
-      const emptyPdfDoc = await PDFDocument.create();
-      const emptyPdfBuffer = await emptyPdfDoc.save();
-      const targetPdfDoc = await PDFDocument.create();
-      targetPdfDoc.addPage();
-      
-      // ACT: Append from the empty source
-      const result = await appendPdfs(emptyPdfBuffer, targetPdfDoc);
-      
-      // ASSERT:
-      expect(result.success).toBe(true);
-      // A "blank" PDF actually has one page when loaded by pdf-lib
-      expect(result.pagesAdded).toBe(1);
-      
-      // Target should now have two pages (original + blank page from empty PDF)
-      expect(targetPdfDoc.getPageIndices().length).toBe(2);
-    });
-  });
-  
+// Helper to extract the content from a PDF for verification
+const extractTextContent = async (pdfBuffer) => {
+  // In a real implementation, we'd use a PDF text extraction tool
+  // For this test, we'll just check if the buffer contains our identifiers
+  const content = pdfBuffer.toString();
+  return content;
+};
+
+describe('PDF Utility Integration Tests', () => {
   describe('mergePdfs', () => {
-    it('should correctly merge multiple PDFs using array syntax', async () => {
-      // ARRANGE: Create sample PDFs with different page counts
-      const pdf1Buffer = await createSamplePdf(2); // 2-page PDF
-      const pdf2Buffer = await createSamplePdf(3); // 3-page PDF
-      const pdf3Buffer = await createSamplePdf(1); // 1-page PDF
+    it('should merge multiple PDF buffers and preserve their content', async () => {
+      // ARRANGE: Create test PDF buffers with unique identifiers
+      const pdf1 = await createTestPdfBuffer('TEST1');
+      const pdf2 = await createTestPdfBuffer('TEST2');
+      const pdf3 = await createTestPdfBuffer('TEST3');
       
-      // ACT: Merge the PDFs
-      const mergedPdfBuffer = await mergePdfs([pdf1Buffer, pdf2Buffer, pdf3Buffer]);
+      // ACT: Merge the PDFs using the public API
+      const mergedPdfBuffer = await mergePdfs([pdf1, pdf2, pdf3]);
       
       // ASSERT:
-      // Verify the result is a buffer
+      // 1. Result is a Buffer
       expect(Buffer.isBuffer(mergedPdfBuffer)).toBe(true);
       
-      // Load the merged document and check its page count
-      const mergedDoc = await PDFDocument.load(mergedPdfBuffer);
-      expect(mergedDoc.getPageCount()).toBe(6); // 2 + 3 + 1 = 6 pages
+      // 2. Result contains content from all source PDFs
+      const extractedContent = await extractTextContent(mergedPdfBuffer);
+      expect(extractedContent.includes('TEST1')).toBe(true);
+      expect(extractedContent.includes('TEST2')).toBe(true);
+      expect(extractedContent.includes('TEST3')).toBe(true);
     });
-    
-    it('should correctly merge PDFs using individual arguments', async () => {
-      // ARRANGE: Create sample PDFs
-      const pdf1Buffer = await createSamplePdf(1);
-      const pdf2Buffer = await createSamplePdf(2);
+
+    it('should merge PDFs when provided as individual arguments', async () => {
+      // ARRANGE: Create test PDF buffers
+      const pdf1 = await createTestPdfBuffer('ARG1');
+      const pdf2 = await createTestPdfBuffer('ARG2');
       
-      // ACT: Merge using individual arguments
-      const mergedPdfBuffer = await mergePdfs(pdf1Buffer, pdf2Buffer);
+      // ACT: Merge using individual arguments syntax
+      const mergedPdfBuffer = await mergePdfs(pdf1, pdf2);
       
       // ASSERT:
-      const mergedDoc = await PDFDocument.load(mergedPdfBuffer);
-      expect(mergedDoc.getPageCount()).toBe(3); // 1 + 2 = 3 pages
+      expect(Buffer.isBuffer(mergedPdfBuffer)).toBe(true);
+      
+      const extractedContent = await extractTextContent(mergedPdfBuffer);
+      expect(extractedContent.includes('ARG1')).toBe(true);
+      expect(extractedContent.includes('ARG2')).toBe(true);
     });
-    
-    it('should create an empty PDF when no inputs are provided', async () => {
+
+    it('should create a valid PDF when no inputs are provided', async () => {
       // ACT: Call with no arguments
-      const emptyPdfBuffer = await mergePdfs();
+      const resultBuffer = await mergePdfs();
       
       // ASSERT:
-      const emptyDoc = await PDFDocument.load(emptyPdfBuffer);
-      // pdf-lib treats an initially empty PDF as having 1 page when loaded
-      expect(emptyDoc.getPageCount()).toBe(1);
+      expect(Buffer.isBuffer(resultBuffer)).toBe(true);
+      // The result should be a valid PDF starting with the PDF header
+      expect(resultBuffer.toString().startsWith('%PDF-')).toBe(true);
+    });
+    
+    it('should handle an empty array of buffers', async () => {
+      // ACT: Pass an empty array
+      const resultBuffer = await mergePdfs([]);
+      
+      // ASSERT:
+      expect(Buffer.isBuffer(resultBuffer)).toBe(true);
+      expect(resultBuffer.toString().startsWith('%PDF-')).toBe(true);
     });
   });
-  
-  // Optional: Test with actual PDF files from the filesystem if available
-  // This is commented out as it depends on the environment
-  /*
-  describe('Using actual PDF files from disk', () => {
-    it('should merge actual PDF files', async () => {
-      // Get paths to sample PDFs (assumes they exist in the repo)
-      const samplePath = path.join(__dirname, '../sample-pdfs');
-      const file1Path = path.join(samplePath, 'sample1.pdf');
-      const file2Path = path.join(samplePath, 'sample2.pdf');
+
+  describe('appendPdfs', () => {
+    it('should append pages from source PDF to target PDF', async () => {
+      // ARRANGE: Create test PDF buffers with unique identifiers
+      const sourcePdf = await createTestPdfBuffer('SOURCE');
+      const targetPdf = await createTestPdfBuffer('TARGET');
       
-      // Read the files
-      const file1Buffer = fs.readFileSync(file1Path);
-      const file2Buffer = fs.readFileSync(file2Path);
+      // ACT: Append source to target
+      const result = await appendPdfs(sourcePdf, targetPdf);
       
-      // Merge them
-      const merged = await mergePdfs(file1Buffer, file2Buffer);
+      // ASSERT:
+      expect(result.success).toBe(true);
+      expect(result.pagesAdded).toBeGreaterThan(0);
+      expect(Buffer.isBuffer(result.buffer)).toBe(true);
       
-      // Validate the merge worked
-      const mergedDoc = await PDFDocument.load(merged);
-      expect(mergedDoc.getPageCount()).toBeGreaterThan(0);
+      // Check that the resulting PDF contains content from both PDFs
+      const extractedContent = await extractTextContent(result.buffer);
+      expect(extractedContent.includes('SOURCE')).toBe(true);
+      expect(extractedContent.includes('TARGET')).toBe(true);
+    });
+
+    it('should handle empty source PDF gracefully', async () => {
+      // ARRANGE: Create an empty source PDF and a target PDF
+      const emptyPdf = Buffer.from('%PDF-1.7\ntrailer<</Root 1 0 R>>\n%%EOF');
+      const targetPdf = await createTestPdfBuffer('TARGET');
+      
+      // ACT: Append empty source to target
+      const result = await appendPdfs(emptyPdf, targetPdf);
+      
+      // ASSERT:
+      expect(result.success).toBe(true);
+      // The exact pagesAdded value depends on how pdf-lib handles empty PDFs
+      expect(Buffer.isBuffer(result.buffer)).toBe(true);
     });
   });
-  */
 });
